fix(google): invalidate cached downloads after file updates
- ensure cached file downloads are refreshed when files are updated
- improve error handling for non-ok JSON responses from the API
- add tests for download, upload, and list functionalities to cover new behaviors
