+++ src/shared/lib/googleDrive/api/simplifiedAPI.listCache.test.ts
@@ -1,30 +1,12 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { HttpStatusCode } from '../../error/httpStatus';
 import { SPACE } from './types';
+import {
+  createDeferred,
+  createFolderMimeType,
+  createJsonResponse,
+} from './simplifiedAPI.testUtils';
 
-const createFolderMimeType = 'application/vnd.google-apps.folder';
-
-const createJsonResponse = (payload: unknown, status = 200) =>
-  new Response(JSON.stringify(payload), {
-    status,
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  });
-
-const createDeferred = <T>() => {
-  let resolve!: (value: T) => void;
-  let reject!: (reason?: unknown) => void;
-
-  const promise = new Promise<T>((res, rej) => {
-    resolve = res;
-    reject = rej;
-  });
-
-  return { promise, resolve, reject };
-};
-
-describe('simplifiedAPI cache invalidation', () => {
+describe('simplifiedAPI list cache invalidation', () => {
   beforeEach(() => {
     vi.resetModules();
   });
@@ -115,7 +97,7 @@ describe('simplifiedAPI cache invalidation', () => {
     ]);
   });
 
-  it('caches empty App Data directory lists', async () => {
+  it('does not cache empty App Data directory lists before create', async () => {
     const fetchMock = vi
       .fn<typeof fetch>()
       .mockResolvedValueOnce(createJsonResponse({ files: [] }))
@@ -648,534 +630,5 @@ describe('simplifiedAPI cache invalidation', () => {
       expect.objectContaining({ id: 'folder-c-id' }),
       expect.objectContaining({ id: 'new-folder-id' }),
     ]);
-  });
-
-  it('fetches only a single page when fetchAll is false', async () => {
-    const fetchMock = vi.fn<typeof fetch>().mockResolvedValue(
-      createJsonResponse({
-        files: [
-          {
-            id: 'folder-a-id',
-            name: 'folder-a',
-            mimeType: createFolderMimeType,
-            modifiedTime: '2024-01-01T00:00:00.000Z',
-            parents: ['parent-id'],
-          },
-        ],
-        nextPageToken: 'next-page',
-      }),
-    );
-
-    vi.stubGlobal('fetch', fetchMock);
-
-    const { clearCaches, getGFileMetaList } = await import('./simplifiedAPI');
-
-    clearCaches();
-
-    const result = await getGFileMetaList(
-      { ACCESS_TOKEN: 'token' },
-      {
-        q: {
-          parentId: 'parent-id',
-          trashed: false,
-        },
-        spaces: [SPACE.drive],
-        pageToken: 'current-page',
-        pageSize: 25,
-        fetchAll: false,
-      },
-    );
-
-    expect(fetchMock).toHaveBeenCalledTimes(1);
-    expect(result).toEqual({
-      files: [expect.objectContaining({ id: 'folder-a-id' })],
-      nextPageToken: 'next-page',
-    });
-  });
-
-  it('rejects create requests with an empty parents array', async () => {
-    const { create } = await import('./simplifiedAPI');
-
-    await expect(
-      create(
-        {
-          ACCESS_TOKEN: 'token',
-        },
-        {
-          name: 'orphan-folder',
-          parents: [],
-          mimeType: createFolderMimeType,
-        },
-      ),
-    ).rejects.toMatchObject({
-      code: HttpStatusCode.FORBIDDEN,
-    });
-  });
-
-  it('reuses the cached file download when metadata has not changed', async () => {
-    const mediaBlob = new Blob(['cached-content'], { type: 'text/plain' });
-    const fetchMock = vi
-      .fn<typeof fetch>()
-      .mockResolvedValueOnce(
-        createJsonResponse({
-          id: 'file-id',
-          name: 'notes.txt',
-          mimeType: 'text/plain',
-          modifiedTime: '2024-01-01T00:00:00.000Z',
-          parents: ['parent-id'],
-        }),
-      )
-      .mockResolvedValueOnce(new Response(mediaBlob, { status: 200 }));
-
-    vi.stubGlobal('fetch', fetchMock);
-
-    const { clearCaches, download } = await import('./simplifiedAPI');
-
-    clearCaches();
-
-    const firstFile = await download({ ACCESS_TOKEN: 'token' }, 'file-id');
-    const secondFile = await download({ ACCESS_TOKEN: 'token' }, 'file-id');
-
-    expect(fetchMock).toHaveBeenCalledTimes(2);
-    expect(await firstFile.text()).toBe('cached-content');
-    expect(secondFile).toBe(firstFile);
-  });
-
-  it('normalizes non-ok JSON responses into GoogleDriveError', async () => {
-    const fetchMock = vi.fn<typeof fetch>().mockResolvedValue(
-      createJsonResponse(
-        {
-          error: {
-            code: HttpStatusCode.FORBIDDEN,
-            message: 'Forbidden',
-          },
-        },
-        HttpStatusCode.FORBIDDEN,
-      ),
-    );
-
-    vi.stubGlobal('fetch', fetchMock);
-
-    const { clearCaches, getGFileMetaList } = await import('./simplifiedAPI');
-
-    clearCaches();
-
-    await expect(
-      getGFileMetaList(
-        { ACCESS_TOKEN: 'token' },
-        {
-          q: {
-            parentId: 'parent-id',
-            trashed: false,
-          },
-          spaces: [SPACE.drive],
-          fetchAll: true,
-        },
-      ),
-    ).rejects.toMatchObject({
-      name: 'GoogleDriveError',
-      code: HttpStatusCode.FORBIDDEN,
-      message: 'Forbidden',
-    });
-  });
-
-  it('uploads string content as text/plain and invalidates cached downloads', async () => {
-    const fetchMock = vi
-      .fn<typeof fetch>()
-      .mockResolvedValueOnce(
-        createJsonResponse({
-          id: 'file-id',
-          name: 'notes.txt',
-          mimeType: 'text/plain',
-          modifiedTime: '2024-01-01T00:00:00.000Z',
-          parents: ['parent-id'],
-        }),
-      )
-      .mockResolvedValueOnce(new Response(new Blob(['cached']), { status: 200 }))
-      .mockResolvedValueOnce(new Response('{}', { status: 200 }))
-      .mockResolvedValueOnce(
-        createJsonResponse({
-          id: 'file-id',
-          name: 'notes.txt',
-          mimeType: 'text/plain',
-          modifiedTime: '2024-01-02T00:00:00.000Z',
-          parents: ['parent-id'],
-        }),
-      )
-      .mockResolvedValueOnce(new Response(new Blob(['fresh']), { status: 200 }));
-
-    vi.stubGlobal('fetch', fetchMock);
-
-    const { clearCaches, download, upload } = await import('./simplifiedAPI');
-
-    clearCaches();
-
-    await download({ ACCESS_TOKEN: 'token' }, 'file-id');
-    await upload({ ACCESS_TOKEN: 'token' }, 'file-id', 'next content');
-    const refreshed = await download({ ACCESS_TOKEN: 'token' }, 'file-id');
-
-    const uploadRequest = fetchMock.mock.calls[2]?.[0];
-
-    expect(uploadRequest).toBeInstanceOf(Request);
-    if (!(uploadRequest instanceof Request)) {
-      throw new Error('Expected upload request to be a Request instance');
-    }
-
-    expect(uploadRequest.method).toBe('PATCH');
-    expect(uploadRequest.headers.get('Authorization')).toBe('Bearer token');
-    expect(uploadRequest.headers.get('Content-Type')).toBe('text/plain');
-    expect(await refreshed.text()).toBe('fresh');
-  });
-
-  it('rejects unsupported upload payloads', async () => {
-    const { upload } = await import('./simplifiedAPI');
-    const unsupportedChunk: WriteParams = {
-      type: 'write',
-      data: null,
-    };
-
-    await expect(upload({ ACCESS_TOKEN: 'token' }, 'file-id', unsupportedChunk)).rejects.toThrow(
-      'Unsupported file type',
-    );
-  });
-});
-
-describe('createWithContent', () => {
-  beforeEach(() => {
-    vi.resetModules();
-  });
-
-  afterEach(() => {
-    vi.restoreAllMocks();
-    vi.unstubAllGlobals();
-  });
-
-  it('sends multipart request with metadata and content', async () => {
-    let requestBody = '';
-    const fetchMock = vi.fn<typeof fetch>().mockImplementation(async (input) => {
-      if (!(input instanceof Request)) {
-        throw new Error('Expected fetch to receive a Request');
-      }
-      requestBody = await input.clone().text();
-
-      return createJsonResponse({
-        id: 'file-id',
-        name: 'notes.txt',
-        mimeType: 'text/plain',
-        size: '7',
-        modifiedTime: '2024-01-02T00:00:00.000Z',
-        parents: ['parent-id'],
-      });
-    });
-    vi.stubGlobal('fetch', fetchMock);
-
-    const { createWithContent } = await import('./simplifiedAPI');
-
-    await createWithContent(
-      { ACCESS_TOKEN: 'token' },
-      {
-        name: 'notes.txt',
-        parents: ['parent-id'],
-      },
-      'content',
-    );
-
-    expect(fetchMock).toHaveBeenCalledTimes(1);
-    const [request] = fetchMock.mock.calls[0] ?? [];
-
-    expect(request instanceof Request).toBe(true);
-    if (!(request instanceof Request)) {
-      throw new Error('Expected fetch to receive a Request');
-    }
-    expect(request.url).toContain(
-      'https://www.googleapis.com/upload/drive/v3/files?uploadType=multipart',
-    );
-    expect(request.method).toBe('POST');
-    expect(request.headers.get('Authorization')).toBe('Bearer token');
-    expect(request.headers.get('Content-Type')).toContain('multipart/related; boundary=');
-
-    // Content-Length should not be set manually (browser controls this)
-    expect(request.headers.has('Content-Length')).toBe(false);
-
-    expect(requestBody).toContain('Content-Type: application/json; charset=UTF-8');
-    expect(requestBody).toContain(JSON.stringify({ name: 'notes.txt', parents: ['parent-id'] }));
-    expect(requestBody).toContain('Content-Type: text/plain');
-    expect(requestBody).toContain('content');
-  });
-
-  it('uses fieldsGDriveFileMeta in multipart request', async () => {
-    const fetchMock = vi.fn<typeof fetch>().mockResolvedValue(
-      createJsonResponse({
-        id: 'file-id',
-        name: 'notes.txt',
-        mimeType: 'text/plain',
-        modifiedTime: '2024-01-02T00:00:00.000Z',
-      }),
-    );
-    vi.stubGlobal('fetch', fetchMock);
-
-    const { createWithContent } = await import('./simplifiedAPI');
-    const { fieldsGDriveFileMeta } = await import('./types');
-
-    await createWithContent(
-      { ACCESS_TOKEN: 'token' },
-      {
-        name: 'notes.txt',
-        parents: ['parent-id'],
-      },
-      new Blob(['content'], { type: 'text/plain' }),
-    );
-
-    const [request] = fetchMock.mock.calls[0] ?? [];
-
-    expect(request instanceof Request).toBe(true);
-    if (!(request instanceof Request)) {
-      throw new Error('Expected fetch to receive a Request');
-    }
-    const requestUrl = new URL(request.url);
-
-    expect(requestUrl.searchParams.get('fields')).toBe(fieldsGDriveFileMeta);
-  });
-
-  it('caches created file metadata in gFileMetaCache', async () => {
-    const fetchMock = vi.fn<typeof fetch>().mockResolvedValue(
-      createJsonResponse({
-        id: 'new-file-id',
-        name: 'notes.txt',
-        mimeType: 'text/plain',
-        size: '7',
-        modifiedTime: '2024-01-02T00:00:00.000Z',
-        parents: ['parent-id'],
-      }),
-    );
-    vi.stubGlobal('fetch', fetchMock);
-
-    const { clearCaches, createWithContent, getGDriveFileMeta } = await import('./simplifiedAPI');
-
-    clearCaches();
-
-    await createWithContent(
-      { ACCESS_TOKEN: 'token' },
-      {
-        name: 'notes.txt',
-        parents: ['parent-id'],
-      },
-      'content',
-    );
-
-    expect(fetchMock).toHaveBeenCalledTimes(1);
-
-    const metadata = await getGDriveFileMeta({ ACCESS_TOKEN: 'token' }, 'new-file-id');
-
-    expect(fetchMock).toHaveBeenCalledTimes(1);
-    expect(metadata).toEqual(
-      expect.objectContaining({
-        id: 'new-file-id',
-        name: 'notes.txt',
-        mimeType: 'text/plain',
-      }),
-    );
-  });
-
-  it('invalidates parent list-cache after createWithContent', async () => {
-    const fetchMock = vi
-      .fn<typeof fetch>()
-      .mockResolvedValueOnce(
-        createJsonResponse({
-          files: [
-            {
-              id: 'existing-file',
-              name: 'existing.txt',
-              mimeType: 'text/plain',
-              modifiedTime: '2024-01-01T00:00:00.000Z',
-              parents: ['parent-id'],
-            },
-          ],
-        }),
-      )
-      .mockResolvedValueOnce(
-        createJsonResponse({
-          id: 'new-file-id',
-          name: 'notes.txt',
-          mimeType: 'text/plain',
-          size: '7',
-          modifiedTime: '2024-01-02T00:00:00.000Z',
-          parents: ['parent-id'],
-        }),
-      )
-      .mockResolvedValueOnce(
-        createJsonResponse({
-          files: [
-            {
-              id: 'existing-file',
-              name: 'existing.txt',
-              mimeType: 'text/plain',
-              modifiedTime: '2024-01-01T00:00:00.000Z',
-              parents: ['parent-id'],
-            },
-            {
-              id: 'new-file-id',
-              name: 'notes.txt',
-              mimeType: 'text/plain',
-              size: '7',
-              modifiedTime: '2024-01-02T00:00:00.000Z',
-              parents: ['parent-id'],
-            },
-          ],
-        }),
-      );
-
-    vi.stubGlobal('fetch', fetchMock);
-
-    const { clearCaches, createWithContent, getGFileMetaList } = await import('./simplifiedAPI');
-
-    clearCaches();
-
-    const auth = { ACCESS_TOKEN: 'token' };
-    const listParams = {
-      q: {
-        parentId: 'parent-id',
-        trashed: false,
-      },
-      spaces: [SPACE.drive],
-      fetchAll: true,
-    };
-
-    await getGFileMetaList(auth, listParams);
-    await createWithContent(
-      auth,
-      {
-        name: 'notes.txt',
-        parents: ['parent-id'],
-      },
-      'content',
-    );
-    const refreshedResult = await getGFileMetaList(auth, listParams);
-
-    expect(fetchMock).toHaveBeenCalledTimes(3);
-    expect(refreshedResult.files).toEqual([
-      expect.objectContaining({ id: 'existing-file' }),
-      expect.objectContaining({ id: 'new-file-id' }),
-    ]);
-  });
-
-  it('negative cache does not hide created file after parent cache invalidation', async () => {
-    const fetchMock = vi
-      .fn<typeof fetch>()
-      .mockResolvedValueOnce(createJsonResponse({ files: [] }))
-      .mockResolvedValueOnce(
-        createJsonResponse({
-          id: 'new-file-id',
-          name: 'notes.txt',
-          mimeType: 'text/plain',
-          size: '7',
-          modifiedTime: '2024-01-02T00:00:00.000Z',
-          parents: ['parent-id'],
-        }),
-      )
-      .mockResolvedValueOnce(
-        createJsonResponse({
-          files: [
-            {
-              id: 'new-file-id',
-              name: 'notes.txt',
-              mimeType: 'text/plain',
-              size: '7',
-              modifiedTime: '2024-01-02T00:00:00.000Z',
-              parents: ['parent-id'],
-            },
-          ],
-        }),
-      );
-
-    vi.stubGlobal('fetch', fetchMock);
-
-    const { clearCaches, createWithContent, getGFileMetaList } = await import('./simplifiedAPI');
-
-    clearCaches();
-
-    const auth = { ACCESS_TOKEN: 'token' };
-    const listParams = {
-      q: {
-        parentId: 'parent-id',
-        trashed: false,
-      },
-      spaces: [SPACE.drive],
-      fetchAll: true,
-    };
-
-    await getGFileMetaList(auth, listParams);
-    await createWithContent(
-      auth,
-      {
-        name: 'notes.txt',
-        parents: ['parent-id'],
-      },
-      'content',
-    );
-    const refreshedResult = await getGFileMetaList(auth, listParams);
-
-    expect(fetchMock).toHaveBeenCalledTimes(3);
-    expect(refreshedResult.files).toEqual([
-      expect.objectContaining({
-        id: 'new-file-id',
-        name: 'notes.txt',
-      }),
-    ]);
-  });
-
-  it('does not cache error responses', async () => {
-    const fetchMock = vi
-      .fn<typeof fetch>()
-      .mockResolvedValueOnce(
-        createJsonResponse(
-          {
-            error: {
-              code: HttpStatusCode.NOT_FOUND,
-              message: 'File not found',
-            },
-          },
-          HttpStatusCode.NOT_FOUND,
-        ),
-      )
-      .mockResolvedValueOnce(
-        createJsonResponse({
-          files: [
-            {
-              id: 'existing-file',
-              name: 'existing.txt',
-              mimeType: 'text/plain',
-              modifiedTime: '2024-01-01T00:00:00.000Z',
-              parents: ['parent-id'],
-            },
-          ],
-        }),
-      );
-
-    vi.stubGlobal('fetch', fetchMock);
-
-    const { clearCaches, getGFileMetaList } = await import('./simplifiedAPI');
-
-    clearCaches();
-
-    const auth = { ACCESS_TOKEN: 'token' };
-    const listParams = {
-      q: {
-        parentId: 'parent-id',
-        trashed: false,
-      },
-      spaces: [SPACE.drive],
-      fetchAll: true,
-    };
-
-    await expect(getGFileMetaList(auth, listParams)).rejects.toMatchObject({
-      name: 'GoogleDriveError',
-      code: HttpStatusCode.NOT_FOUND,
-    });
-
-    const result = await getGFileMetaList(auth, listParams);
-
-    expect(fetchMock).toHaveBeenCalledTimes(2);
-    expect(result.files).toEqual([expect.objectContaining({ id: 'existing-file' })]);
   });
 });
